Added overwritten methods for 'query'.
